feat: Added helpers for server-side rendering
Added methods to serialize and hydrate the state of the router.

BREAKING CHANGE: The `RouterStore` constructor has a small change in its arguments - the last
argument `initialRoute` is replaced by `initialState`. This argument was seldom used, so it should
not affect most users. The new signature now supports hydration from the state provided by the
server.